RunLoop::isMain() should not need to do any heap allocations https://bugs.webkit.org/show_bug.cgi?id=223227
Reviewed by Darin Adler.

RunLoop::isMain() should not need to do any heap allocations. Before this change,
calling RunLoop::isMain() on a non-main thread would call RunLoop::current() which
would allocate the RunLoop for the current thread. This is inefficient and an issue
for WebAudio since we're not allowed to do heap allocation on the audio rendering
thread.

* wtf/RunLoop.cpp:
(WTF::RunLoop::runLoopHolder):
(WTF::RunLoop::current):
(WTF::RunLoop::isMain):
* wtf/RunLoop.h:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274458 268f45cc-cd09-0410-ab3c-d52691b4dbfc