[Cocoa] YouTube audio stalls occasionally https://bugs.webkit.org/show_bug.cgi?id=223233 <rdar://75057532>
Reviewed by Eric Carlson.

In r273513, we added support for parsing of audio frames which spanned
appends. But in so doing, left in a bug where the size and starting byte
offset of those audio frames was miscalculated. Previously, the starting
byte offset was reset during each pass through consumeFrameData(), and the
size of the m_packetData was increased by the full packet size each pass
as well, leading to both an incorrect starting point for the frame and an
incorrect overall size of the combined packet data. This lead to a decoding
error when the parser hit the empty frame bytes after the last successful
partial parse.

Now, we will add explicit tracking of the current frame's starting byte
offset as well as the frame's size, reset after each successful full frame
parse.

* platform/graphics/cocoa/SourceBufferParserWebM.cpp:
(WebCore::SourceBufferParserWebM::AudioTrackData::consumeFrameData):
* platform/graphics/cocoa/SourceBufferParserWebM.h:


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274485 268f45cc-cd09-0410-ab3c-d52691b4dbfc