REGRESSION (r272300): [iOS] ASSERTION FAILED: Unsafe to ref/deref from different threads under WebViewLayerFlushScheduler::layerFlushCallback https://bugs.webkit.org/show_bug.cgi?id=222712 <rdar://problem/75022845>
Reviewed by Simon Fraser.

We are scheduling a layer flush from a layerFlushCallback and so rescheduling the RunLoopObserver.
This is fine in itself but becomes a problem when the main thread web lock auto-unlock observer
runs before that observer. Then that layer flush runs without the web lock held and messes up
ref/deref main thread asserts (isMainThread() returns false when the web lock is not held).

* WebView/WebViewData.mm:
(WebViewLayerFlushScheduler::layerFlushCallback):

Fix by explicitly taking the lock.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274500 268f45cc-cd09-0410-ab3c-d52691b4dbfc