Avoid heap allocation under AudioDestinationCocoa::render() https://bugs.webkit.org/show_bug.cgi?id=223272
Reviewed by Sam Weinig.

In the common case where no AudioWorklet is used, AudioDestinationCocoa::render()
would unnecessarily create a WTF::Function to dispatch to the render thread, even
though we are already on the render thread. Constructing a WTF::Function does a
heap allocation so we really want to avoid constructing one on the high priority
audio thread.

In the AudioWorklet case, we still construct a WTF::Function for now. I will follow
up if I find a way to avoid it in this case too.

* Modules/webaudio/DefaultAudioDestinationNode.cpp:
(WebCore::Function<void):
* platform/audio/cocoa/AudioDestinationCocoa.cpp:
(WebCore::AudioDestinationCocoa::render):
(WebCore::AudioDestinationCocoa::renderOnRenderingTheadIfPlaying):
* platform/audio/cocoa/AudioDestinationCocoa.h:
* platform/audio/gstreamer/WebKitWebAudioSourceGStreamer.cpp:
(webKitWebAudioSrcRenderIteration):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274527 268f45cc-cd09-0410-ab3c-d52691b4dbfc