REGRESSION (r272019): Momentum scrolling in proximity css-scroll-snap scroll areas broken https://bugs.webkit.org/show_bug.cgi?id=223375
Patch by Martin Robinson <mrobinson@igalia.com> on 2021-03-17
Reviewed by Simon Fraser.

Selectively revert the problematic refactoring in r272019. It is difficult to revert
the entire change since a non-trivial amount of subsequent changes depend on it.

No new tests. This change is a selective revert and animation for momentum
scrolling is very difficult to test.

* platform/ScrollSnapAnimatorState.cpp:
(WebCore::ScrollSnapAnimatorState::targetOffsetForStartOffset const): Revert changes
from r272019.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274564 268f45cc-cd09-0410-ab3c-d52691b4dbfc