RTCRtpScriptTransformer::writable should check for m_backend https://bugs.webkit.org/show_bug.cgi?id=223373
Reviewed by Eric Carlson.

Covered by existing tests.

* Modules/mediastream/RTCRtpScriptTransformer.cpp:
(WebCore::RTCRtpScriptTransformer::writable):
The writer can only start writing after readable is started, which means m_backend is set.
But the transformer may be cleared at which point m_backend is nullified but there might be some encoded frames in the pipe to be written.
Add a null check to fix this.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274623 268f45cc-cd09-0410-ab3c-d52691b4dbfc