Pagination mode: Increase the amount of recursions allowed by RenderTable::layout() https://bugs.webkit.org/show_bug.cgi?id=218575
Patch by Carlos Garcia Campos <cgarcia@igalia.com> on 2021-03-18
Reviewed by Zalan Bujtas.

In r219394 a recursion check was introduced to avoid infinite recursion when a section is always moved on
RenderTable::layout(). But sometimes it's needed to recurse more than once to layout the table when section move
due to the pagination. This patch limits the recursion level to the amount of sections in the table, instead of
just one.

* rendering/RenderTable.cpp:
(WebCore::RenderTable::layout):
* rendering/RenderTable.h:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274627 268f45cc-cd09-0410-ab3c-d52691b4dbfc