Avoid heap allocation under AudioWorkletNode::process() https://bugs.webkit.org/show_bug.cgi?id=223452
Reviewed by Darin Adler.

Always use arrays of size 128 (render quantum) for AudioParam values. We used to
dynamically choose between a size of 128 and 1 (when there is no automation
scheduled and the value is constant for the quantum). This avoids having to make
heap allocations on the audio thread, which are bad for performance.

Content (and existing AudioWorklet tests) can handle either size because the
specification says [1]:
"""
For each array, the array contains the computedValue of the parameter for all
frames in the render quantum. However, if no automation is scheduled during this
render quantum, the array MAY have length 1 with the array element being the
constant value of the AudioParam for the render quantum.
"""

Note that the specification uses "MAY" so this is not a requirement for us to use
an array of size 1 in such case. I also verified that Blink is never using an
array of size 1 either. They always use arrays of 128 (render quantum size) and
std::fill() them with the constant value when there is no automation. Our new
behavior is thus consistent with Blink.

[1] https://webaudio.github.io/web-audio-api/#audioworkletprocess-callback-parameters

* Modules/webaudio/AudioWorkletNode.cpp:
(WebCore::AudioWorkletNode::initializeAudioParameters):
(WebCore::AudioWorkletNode::process):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274677 268f45cc-cd09-0410-ab3c-d52691b4dbfc