Add a fast path for subdivideAndResolveStyle() https://bugs.webkit.org/show_bug.cgi?id=223480
Reviewed by Myles Maxfield.

Most of the time, InlineTextBox::paint() will be painting text runs with no marked text,
so optimize subdivideAndResolveStyle() for this case.

This removes 1.6% of the time in painting in the MotionMark Design subtest.

* rendering/MarkedTextStyle.cpp:
(WebCore::subdivideAndResolveStyle):
(WebCore::coalesceAdjacentMarkedTexts): No need to do any work if there are zero or one
ranges.


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274700 268f45cc-cd09-0410-ab3c-d52691b4dbfc