Async Clipboard read prevents WebRTC IOSurfaces from being released https://bugs.webkit.org/show_bug.cgi?id=223489 <rdar://problem/75601433>
Reviewed by Eric Carlson.

Instead of dispatching a task to main thread, the task refing a MediaSample,
we store the MediaSample in MediaPlayerPrivateMediaStreamAVFObjC on the background thread and
access it on main thread with a lock.
This ensures that we do not keep more than one media sample per MediaPlayerPrivateMediaStreamAVFObjC, even if main thread is blocked.
Make sure to reset the video transformation according hasChangedOrientation even if there is no media sample available anymore.
Covered by existing tests.

* platform/graphics/avfoundation/objc/MediaPlayerPrivateMediaStreamAVFObjC.h:
* platform/graphics/avfoundation/objc/MediaPlayerPrivateMediaStreamAVFObjC.mm:
(WebCore::MediaPlayerPrivateMediaStreamAVFObjC::processNewVideoSample):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274778 268f45cc-cd09-0410-ab3c-d52691b4dbfc