[JSC] Intl.Locale should not assume is8Bit https://bugs.webkit.org/show_bug.cgi?id=223553
Reviewed by Ross Kirsling.

JSTests:

* stress/intl-locale-non-8bit.js: Added.
(shouldBe):

Source/JavaScriptCore:

is8Bit or not is not guaranteed if it is an user-input. For example, "test日本語".substring(0, 3) should be non 8Bit string.
Intl.Locale has several places that assumed that input should be 8Bit if they are ASCII. This patch fixes it.

* runtime/IntlLocale.cpp:
(JSC::LocaleIDBuilder::overrideLanguageScriptRegion):
(JSC::LocaleIDBuilder::setKeywordValue):

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274784 268f45cc-cd09-0410-ab3c-d52691b4dbfc