[Web Animations] nullptr crash in updatePlaybackRate() with no timeline https://bugs.webkit.org/show_bug.cgi?id=223095
Patch by Ian Gilbert <iang@apple.com> on 2021-03-22
Reviewed by Ryosuke Niwa.

Changed to WebAnimation::remove() to use setTimeline() instead of setTimelineInternal()
to ensure the animation is in the correct play state.

No new tests as this relies on navigating while updating the playback rate and I was unable to make
a reliable test case.

* animation/WebAnimation.cpp:
(WebCore::WebAnimation::remove): Changed setTimelineInternal() to setTimeline() in order to
ensure the animation has the correct play state.

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274824 268f45cc-cd09-0410-ab3c-d52691b4dbfc