wtf/text/IntegerToStringConversion.h:54:104: runtime error: negation of -2147483648 cannot be represented in type 'int'; cast to an unsigned type to negate this value to itself https://bugs.webkit.org/show_bug.cgi?id=223614
Reviewed by Darin Adler.

Source/WTF:

numberToStringSigned() was calling -number when number is negative. This could trigger undefined behavior
since -number may not be representable in the input type. For example, if the type was int and number was
INT_MIN, -INT_MIN would overflow and the result would be undefined. To address the issue, we now cast
number to the destination *unsigned* type *before* negating it.

* wtf/text/IntegerToStringConversion.h:
(WTF::numberToStringSigned):

Tools:

Add API test coverage.

* TestWebKitAPI/Tests/WTF/StringBuilder.cpp:
(TestWebKitAPI::TEST):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@274878 268f45cc-cd09-0410-ab3c-d52691b4dbfc