Drop unnecessary Settings pre-warming in ProcessWarming::prewarmGlobally() https://bugs.webkit.org/show_bug.cgi?id=223749
Reviewed by Antti Koivisto.

Drop unnecessary Settings pre-warming in ProcessWarming::prewarmGlobally(). Based on the comment,
it looks like the intention was to prewarm some font families. However, looking at the code, nothing
that gets populated is global and survives the destruction of the Settings object on the next line.
We don't even make any calls to CTFont.

* page/ProcessWarming.cpp:
(WebCore::ProcessWarming::prewarmGlobally):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@275045 268f45cc-cd09-0410-ab3c-d52691b4dbfc