It should be possible to drag images with overlay content https://bugs.webkit.org/show_bug.cgi?id=223766 <rdar://problem/75856030>
Reviewed by Tim Horton.

Source/WebCore:

Make a few minor adjustments to make drag and drop play well with images with overlay content. See below for
more details.

Test: DragAndDropTests.DragElementWithImageOverlay

* html/HTMLElement.cpp:
(WebCore::HTMLElement::isInsideImageOverlay):

Add a helper method to return whether an individual DOM node is inside an image overlay. Use this in the
existing version of this method that takes a `SimpleRange`.

* html/HTMLElement.h:
* page/DragController.cpp:
(WebCore::DragController::draggableElement const):

Skip all drag source actions except for `Selection` in the case where we're hit-testing to text inside of an
image overlay.

* page/EventHandler.cpp:
(WebCore::EventHandler::handleMousePressEvent):

Make `isMouseDownOnLinkOrImage` `false` in the case where we're over image overlay text.

* rendering/HitTestResult.cpp:
(WebCore::HitTestResult::nodeForImageData const):
(WebCore::HitTestResult::image const):
(WebCore::HitTestResult::imageRect const):
(WebCore::HitTestResult::absoluteImageURL const):

Teach `HitTestResult` to skip out of image overlay content and find the actual image element when determining
the image URL, image, or image rect.

* rendering/HitTestResult.h:

Tools:

Add an API test to verify that the top part of the image (with an overlay) doesn't trigger a `dragstart` when
dragged, but the bottom part of the image does.

* TestWebKitAPI/TestWebKitAPI.xcodeproj/project.pbxproj:
* TestWebKitAPI/Tests/WebKit/simple-image-overlay.html: Added.

Add a test page that installs an image overlay for testing purposes (assuming that the web view injects the
global `internals` object).

* TestWebKitAPI/Tests/WebKitCocoa/DragAndDropTests.mm:
* TestWebKitAPI/Tests/mac/DragAndDropTestsMac.mm:
* TestWebKitAPI/cocoa/DragAndDropSimulator.h:
* TestWebKitAPI/ios/DragAndDropSimulatorIOS.mm:
(-[DragAndDropSimulator containsDraggedType:]):

Add a helper method to return whether or not the simulated drag resulted in the given type being written to the
"drag pasteboard" (in the case of iOS, this just means any of the source item providers).

* TestWebKitAPI/mac/DragAndDropSimulatorMac.mm:
(-[DragAndDropSimulator initWithWebViewFrame:configuration:]):

Make a minor adjustment to ensure that the drag pasteboard is cleared out before starting a simulated drag.

(-[DragAndDropSimulator containsDraggedType:]):


git-svn-id: http://svn.webkit.org/repository/webkit/trunk@275072 268f45cc-cd09-0410-ab3c-d52691b4dbfc